Extract class that can be used for testing other services
